+++ catwalk/models/rank_classification.py
@@ -1,5 +1,6 @@
 import collections
-from typing import Dict, Any, List, Tuple, Sequence, Iterator, Union, Mapping, Optional, cast, Callable
+from dataclasses import dataclass
+from typing import Dict, Any, List, OrderedDict, Tuple, Sequence, Iterator, Union, Mapping, Optional, cast, Callable
 
 import more_itertools
 import torch
@@ -8,11 +9,12 @@ from tango.integrations.torch.util import resolve_device
 from torch import log_softmax
 from torch.nn.utils.rnn import pad_sequence
 from transformers import AutoModelForCausalLM, AutoModelForSeq2SeqLM, T5ForConditionalGeneration, GPT2LMHeadModel, \
-    AutoTokenizer, GPT2Tokenizer, T5TokenizerFast
+    AutoTokenizer, GPT2Tokenizer, T5TokenizerFast, BatchEncoding
 
 from catwalk import cached_transformers
 from catwalk.model import Model, TrainableModel, Instance
 from catwalk.task import Task, InstanceFormat, RankClassificationInstance
+from catwalk.utils import PrefixTrie
 
 _Model = Union[T5ForConditionalGeneration, GPT2LMHeadModel]
 _Tokenizer = Union[T5TokenizerFast, GPT2Tokenizer]
@@ -26,6 +28,7 @@ class RankClassificationModel(Model):
         pretrained_model_name_or_path: str,
         *,
         likelihood_averaging: str = 'char',
+        override_weights_file: str = None,
         **model_kwargs
     ):
         """
@@ -36,16 +39,20 @@ class RankClassificationModel(Model):
         likelihood_averaging : `str`, optional (default = `char`)
             The method for averaging the sum likelihood of the continuation. 'char' averages by 
             character length, 'token' averages by token length.
+        override_weights_file : `str`, optional (default = `None`)
+            If set, this specifies a file from which to load alternate weights that override the
+            weights from huggingface. The file is expected to contain a PyTorch `state_dict`, created
+            with `torch.save()`.
         model_kwargs:
             Additional kwargs passed to the `_make_model` method.
         """
         assert likelihood_averaging in {'char', 'token'}
         self.pretrained_model_name_or_path = pretrained_model_name_or_path
         self.likelihood_averaging = likelihood_averaging
+        self.override_weights_file = override_weights_file
         self.model_kwargs = model_kwargs
 
-    @classmethod
-    def _make_model(cls, pretrained_model_name_or_path: str, **kwargs) -> _Model:
+    def _make_model(cls, pretrained_model_name_or_path: str, *, override_weights_file: str = None, **kwargs) -> _Model:
         raise NotImplementedError
 
     def predict(  # type: ignore
@@ -59,7 +66,7 @@ class RankClassificationModel(Model):
         fewshot_seed: int = None
     ) -> Iterator[Dict[str, Any]]:
         device = resolve_device()
-        model = self._make_model(self.pretrained_model_name_or_path, **self.model_kwargs).to(device).eval()
+        model = self._make_model(self.pretrained_model_name_or_path, override_weights_file=self.override_weights_file, **self.model_kwargs).to(device).eval()
         tokenizer = cached_transformers.get_tokenizer(AutoTokenizer, self.pretrained_model_name_or_path)
 
         for instance_chunk in more_itertools.chunked(instances, max_instances_in_memory):
@@ -192,8 +199,8 @@ class TrainableRankClassificationModel(TrainableModel):
 @Model.register("rc::encoder_decoder")
 class EncoderDecoderRCModel(RankClassificationModel):
     @classmethod
-    def _make_model(cls, pretrained_model_name_or_path: str, **kwargs) -> T5ForConditionalGeneration:
-        return cached_transformers.get(AutoModelForSeq2SeqLM, pretrained_model_name_or_path, False)
+    def _make_model(cls, pretrained_model_name_or_path: str, *, override_weights_file: str = None, **kwargs) -> T5ForConditionalGeneration:
+        return cached_transformers.get(AutoModelForSeq2SeqLM, pretrained_model_name_or_path, False, override_weights_file=override_weights_file)
 
     def _run_loglikelihood(
         self,
@@ -242,11 +249,11 @@ class EncoderDecoderRCModel(RankClassificationModel):
                     for field_name, tensors in unpadded_batch.items()
                 }
 
-                batch_logits = log_softmax(model(**padded_batch).logits, dim=-1).cpu()
+                batch_logits = log_softmax(model(**padded_batch).logits, dim=-1)
 
                 for i, instance_logits, decoder_input_ids in zip(batch_of_indices, batch_logits, unpadded_batch["labels"]):
                     instance_logits = instance_logits[:len(decoder_input_ids)]
-                    instance_logits = torch.gather(instance_logits, 1, decoder_input_ids.unsqueeze(-1))
+                    instance_logits = torch.gather(instance_logits, 1, decoder_input_ids.unsqueeze(-1).to(model.device))
                     denom = len(tuples[i][1]) if self.likelihood_averaging == 'char' else len(decoder_input_ids)
                     results[i] = float(instance_logits.sum()) / denom
 
@@ -254,11 +261,48 @@ class EncoderDecoderRCModel(RankClassificationModel):
         return cast(Sequence[float], results)
 
 
+@dataclass
+class CacheData:
+    cached_sequence: Optional[Sequence[Optional[int]]] = None
+    cached_past_key_values: torch.Tensor = None
+    longest_prefix_to_indices: Optional[Dict[Sequence[Optional[int]],Sequence[int]]] = None
+    indices_to_longest_prefix: Optional[OrderedDict[int,Sequence[Optional[int]]]] = None
+
 @Model.register("rc::decoder_only")
 class DecoderOnlyRCModel(RankClassificationModel):
+    def __init__(
+        self,
+        pretrained_model_name_or_path: str,
+        *,
+        likelihood_averaging: str = 'char',
+        override_weights_file: str = None,
+        prefix_caching: bool = False,
+        **model_kwargs
+    ):
+        """
+        # Parameters
+
+        pretrained_model_name_or_path : `str`
+            The name of the transformer, for example `"gpt2-large"`
+        likelihood_averaging : `str`, optional (default = `char`)
+            The method for averaging the sum likelihood of the continuation. 'char' averages by 
+            character length, 'token' averages by token length.
+        override_weights_file : `str`, optional (default = `None`)
+            If set, this specifies a file from which to load alternate weights that override the
+            weights from huggingface. The file is expected to contain a PyTorch `state_dict`, created
+            with `torch.save()`.
+        prefix_caching : `bool`, optional (default = `False`)
+            If set to True uses a caching strategy that improves performance when many inputs in a task 
+            share prefixes. This orders the dataset by common prefixes and caches the current shared prefix.
+        model_kwargs:
+            Additional kwargs passed to the `_make_model` method.
+        """
+        super().__init__(pretrained_model_name_or_path, likelihood_averaging=likelihood_averaging, override_weights_file=override_weights_file, **model_kwargs)
+        self.prefix_caching = prefix_caching
+
     @classmethod
-    def _make_model(cls, pretrained_model_name_or_path: str, **kwargs) -> GPT2LMHeadModel:
-        return cached_transformers.get(AutoModelForCausalLM, pretrained_model_name_or_path, False)
+    def _make_model(cls, pretrained_model_name_or_path: str, *, override_weights_file: str = None, **kwargs) -> GPT2LMHeadModel:
+        return cached_transformers.get(AutoModelForCausalLM, pretrained_model_name_or_path, False, override_weights_file=override_weights_file)
 
     def _run_loglikelihood(
         self,
@@ -267,8 +311,18 @@ class DecoderOnlyRCModel(RankClassificationModel):
         tokenizer: _Tokenizer,
         batch_size: int = 32,
     ) -> Sequence[float]:
+        cache = CacheData() if self.prefix_caching else None
+
         tokenized_contexts = tokenizer([t[0] for t in tuples])
         tokenized_continuations = tokenizer([t[1] for t in tuples])
+
+        self._final_truncatation(
+            tokenized_contexts, tokenized_continuations, tokenizer.model_max_length
+        )
+
+        ordered_indices = self._reorder_instances(
+            tokenized_contexts, tokenized_continuations, cache
+        )
 
         # transpose the token ids so we can access them one instance at a time
         cc_pairs: List[Dict[str, Tuple[torch.Tensor, torch.Tensor]]] = []
@@ -287,14 +341,6 @@ class DecoderOnlyRCModel(RankClassificationModel):
                     torch.tensor(continuation, dtype=torch.long)
                 )
 
-        # find out the order to process sequences in
-        lengths = torch.tensor([
-            len(cc_pair["input_ids"][0]) + len(cc_pair["input_ids"][1])
-            for cc_pair in cc_pairs
-        ], dtype=torch.int)
-        ordered_indices = torch.argsort(lengths, descending=True)
-        del lengths
-
         # actually do the processing
         results: List[Optional[float]] = [None] * len(ordered_indices)
         with torch.inference_mode():
@@ -302,32 +348,199 @@ class DecoderOnlyRCModel(RankClassificationModel):
                 Tqdm.tqdm(ordered_indices, desc="Running log-likelihood queries"),
                 batch_size)
             for batch_of_indices in batches_of_indices:
-                unpadded_batch = collections.defaultdict(list)
-                input_lengths = []
-                batch_contexts = []
-                batch_continuations = []
-                for index in batch_of_indices:
-                    for field_name, (context_ids, continuation_ids) in cc_pairs[index].items():
-                        ids = torch.cat([context_ids, continuation_ids])
-                        ids = ids[-(tokenizer.model_max_length+1):][:-1]
-                        unpadded_batch[field_name].append(ids)
-
-                    input_lengths.append(len(unpadded_batch["input_ids"][-1]))
-                    batch_contexts.append(cc_pairs[index]["input_ids"][0])
-                    batch_continuations.append(cc_pairs[index]["input_ids"][1])
-
-                padded_batch = {
-                    field_name: pad_sequence(tensors, batch_first=True).to(model.device)
-                    for field_name, tensors in unpadded_batch.items()
-                }
-
-                batch_logits = log_softmax(model(**padded_batch)[0], dim=-1).cpu()
+                inputs, input_lengths, batch_contexts, batch_continuations = self._get_inputs(batch_of_indices, cc_pairs, model, cache)
+                batch_logits = log_softmax(model(**inputs)[0], dim=-1)
                 z = zip(batch_of_indices, batch_logits, input_lengths, batch_contexts, batch_continuations)
                 for i, instance_logits, input_length, instance_context, instance_continuation in z:
+                    assert input_length-len(instance_continuation) >=0
                     instance_logits = instance_logits[input_length-len(instance_continuation):input_length]
-                    instance_logits = torch.gather(instance_logits, 1, instance_continuation.unsqueeze(-1))
+                    instance_logits = torch.gather(instance_logits, 1, instance_continuation.unsqueeze(-1).to(model.device))
                     denom = len(tuples[i][1]) if self.likelihood_averaging == 'char' else len(instance_continuation)
                     results[i] = float(instance_logits.sum()) / denom
 
         assert None not in results
         return cast(Sequence[float], results)
+
+    def _final_truncatation(self, tokenized_contexts: BatchEncoding, tokenized_continuations: BatchEncoding, model_max_length: int):
+        """ Apply a last pass of truncation on the concatenated inputs to make sure it fits in the model_max_length"""
+        assert len(tokenized_contexts['input_ids']) == len(tokenized_continuations['input_ids'])
+        for i in range(len(tokenized_contexts['input_ids'])):
+            context_len = len(tokenized_contexts['input_ids'][i])
+            cont_len = len(tokenized_continuations['input_ids'][i])
+            assert cont_len < model_max_length
+            if context_len +  cont_len > model_max_length:
+                tokenized_contexts['input_ids'][i] = tokenized_contexts['input_ids'][i][-model_max_length + cont_len:]
+                tokenized_contexts['attention_mask'][i] = tokenized_contexts['attention_mask'][i][-model_max_length + cont_len:]
+    
+    def _reorder_instances(self, tokenized_contexts: BatchEncoding, tokenized_continuations: BatchEncoding, cache: CacheData = None) -> Sequence[int]:
+        if self.prefix_caching:
+            assert cache is not None, 'prefix reordering requires a CacheData object'
+            return self._reorder_by_prefix(tokenized_contexts, tokenized_continuations, cache)
+        else:
+            return self._reorder_by_longest(tokenized_contexts, tokenized_continuations)
+    
+    def _reorder_by_prefix(self, tokenized_contexts: BatchEncoding, tokenized_continuations: BatchEncoding, cache: CacheData) -> Sequence[int]:
+        combined_ids = [context + continuation for context, continuation in zip(tokenized_contexts['input_ids'], tokenized_continuations['input_ids'])]
+        cache.longest_prefix_to_indices = self._greedy_assign_prefix_by_total_coverage(combined_ids)
+        cache.indices_to_longest_prefix = OrderedDict()
+        # secondarily sort by length so that largest batches that may cause memory overflow are likely come early
+        for prefix in sorted(cache.longest_prefix_to_indices.keys(), key = lambda x : -len(x)):
+            # sequence indices for each prefix are already sorted by length from reading trie from leaf to root
+            for index in cache.longest_prefix_to_indices[prefix]:
+                cache.indices_to_longest_prefix[index] = prefix
+        return list(cache.indices_to_longest_prefix.keys())
+    
+    def _greedy_assign_prefix_by_total_coverage(self, sequences: Sequence[Sequence[int]]) -> Dict[Sequence[Optional[int]],Sequence[int]]:
+        """Returns a Dict of prefixes and the sequence indices assigned to them. Sorts possible prefixes by total tokens covered in 
+        subsequences and assigns sequences to the first prefix they appear in. PrefixTrie only tracks subsequences after a minimum 
+        track_after_depth so short coincidental overlaps are be ignored."""
+        longest_prefix_to_indices: Dict[Sequence[Optional[int]],Sequence[int]] = {}
+        trie = PrefixTrie(sequences)
+        leaves = trie.get_leaf_nodes()
+        leaves_sequences = [tuple(leaf.get_sequence()) for leaf in leaves]
+        leaves_and_sequences = Tqdm.tqdm(zip(leaves_sequences, leaves), desc="Finding prefixes", total=len(leaves))
+        leaves2prefixes = {leaf_sequence:leaf.get_subsequences() for leaf_sequence, leaf in leaves_and_sequences}
+
+        # greedily assign sequences to prefixes with top total coverage
+        indices_already_assigned = set()
+        for leaf_sequence in sorted(leaves_sequences, key=lambda leaf_sequence : -leaves2prefixes[leaf_sequence][1]):
+            prefix_indices, _ = leaves2prefixes[leaf_sequence]
+            prefix_indices = [prefix_index for prefix_index in prefix_indices if prefix_index not in indices_already_assigned]
+            indices_already_assigned.update(prefix_indices)
+            if len(prefix_indices) > 0:
+                longest_prefix_to_indices[leaf_sequence] = tuple(prefix_indices)
+
+        return longest_prefix_to_indices
+
+    def _reorder_by_longest(self, tokenized_contexts: BatchEncoding, tokenized_continuations: BatchEncoding) -> Sequence[int]:
+        assert len(tokenized_contexts['input_ids']) == len(tokenized_continuations['input_ids'])
+        lengths = torch.tensor([
+            len(tokenized_contexts["input_ids"][i]) + len(tokenized_continuations["input_ids"][i])
+            for i in range(len(tokenized_contexts['input_ids']))
+        ], dtype=torch.int)
+        return torch.argsort(lengths, descending=True).tolist()
+
+    def _get_inputs(self, batch_of_indices: Sequence[int], cc_pairs: List[Dict[str, Tuple[torch.Tensor, torch.Tensor]]], model: _Model, cache: CacheData = None):
+        if self.prefix_caching:
+            assert cache is not None
+            return self._get_inputs_with_cache(batch_of_indices, cc_pairs, model, cache)
+        else:
+            return self._get_inputs_without_cache(batch_of_indices, cc_pairs, model)
+            
+
+    def _get_inputs_with_cache(self, batch_of_indices: Sequence[int], cc_pairs: List[Dict[str, Tuple[torch.Tensor, torch.Tensor]]], model: _Model, cache: CacheData):
+        assert cache.indices_to_longest_prefix is not None
+        prefixes = [cache.indices_to_longest_prefix[index] for index in batch_of_indices]
+        prefix2cache = OrderedDict()
+
+        # compute prefixes
+        if prefixes[0] == cache.cached_sequence:
+            prefix2cache[prefixes[0]] = cache.cached_past_key_values
+
+        uncached_prefixes = list(set(prefix for prefix in prefixes if prefix not in prefix2cache)) # ordering must be fixed
+        if len(uncached_prefixes) > 0:
+            unpadded_prefixes = [torch.tensor(prefix) for prefix in uncached_prefixes]
+            unpadded_prefix_mask = [torch.ones_like(prefix) for prefix in unpadded_prefixes]
+            padded_prefixes = pad_sequence(unpadded_prefixes, batch_first=True).to(model.device)
+            padded_prefix_masks = pad_sequence(unpadded_prefix_mask, batch_first=True, padding_value=0.0).to(model.device)
+            past_key_values = model(input_ids=padded_prefixes, attention_mask=padded_prefix_masks).past_key_values
+            # tensor(layers, keys/values, batch_size, num_heads, sequence_len, embed_size_per_head)
+            past_key_values = torch.stack(tuple(torch.stack(past_key_values[i]) for i in range(len(past_key_values))))
+            for i, prefix in enumerate(uncached_prefixes):
+                # tensor(layers, keys/values, num_heads, sequence_len, embed_size_per_head)
+                prefix2cache[prefix] = past_key_values[:,:,i,:,:len(prefix),:]
+
+        # update cache with last one retrieved since instances come in order by common prefix
+        cache.cached_sequence = prefixes[-1]
+        cache.cached_past_key_values = prefix2cache[prefixes[-1]]
+
+        # pad and mask batched past_key_values
+        unpadded_past_keys_values = [prefix2cache[prefix] for prefix in prefixes]
+        unpadded_past_keys_values_attn_mask = []
+        # only use the prefixed part of past_key_values that is present in the instance
+        for prefix_idx, cc_pairs_idx in enumerate(batch_of_indices):
+            is_identical = True
+            for tok_idx, tok in enumerate(cc_pairs[cc_pairs_idx]['input_ids'][0]):
+                if tok.item() != prefixes[prefix_idx][tok_idx]:
+                    unpadded_past_keys_values_attn_mask.append(torch.tensor([1] * tok_idx, dtype=torch.int64))
+                    is_identical = False
+                    break
+            if is_identical:
+                # Avoid empty input by leaving last token of context for input because continuations drop one token for right shift
+                max_prefix_len = len(cc_pairs[cc_pairs_idx]['input_ids'][0]) - 1
+                unpadded_past_keys_values_attn_mask.append(torch.tensor([1] * max_prefix_len, dtype=torch.int64))
+        
+        # past_keys_values needs its own attention mask
+        padded_past_keys_values_attn_mask = pad_sequence(unpadded_past_keys_values_attn_mask, batch_first=True, padding_value=0)
+        cache_lengths = [mask.sum().item() for mask in padded_past_keys_values_attn_mask]
+        max_past_key_value_len = max(cache_lengths)
+
+        # pad and truncate past_keys_values to longest actually used
+        unpadded_past_keys_values = [t.transpose(0,-2) for t in unpadded_past_keys_values]
+        padded_past_keys_values = pad_sequence(unpadded_past_keys_values, batch_first=True)
+        padded_past_keys_values = padded_past_keys_values.permute((4, 2, 0, 3, 1, 5))
+        # tensor(layers, keys/values, batch_size, num_heads, sequence_len, embed_size_per_head)
+        padded_past_keys_values = padded_past_keys_values[:,:,:,:,:max_past_key_value_len]
+
+        # make input_ids by removing whatever parts of past_key_values are present
+        unpadded_input_ids = []
+        input_lengths = []
+        batch_contexts = []
+        batch_continuations = []
+
+        for prefix_idx, cc_pairs_idx in enumerate(batch_of_indices):
+            context_ids, continuation_ids = cc_pairs[cc_pairs_idx]['input_ids']
+            ids = torch.cat([context_ids, continuation_ids])[:-1]
+            ids = ids[cache_lengths[prefix_idx]:]
+
+            # used to find logits specifically for continuation
+            input_lengths.append(len(ids))
+            batch_contexts.append(cc_pairs[cc_pairs_idx]["input_ids"][0])
+            batch_continuations.append(cc_pairs[cc_pairs_idx]["input_ids"][1])
+
+            unpadded_input_ids.append(ids)
+
+        # batch and pad and make attention mask
+        unpadded_attn_mask = [torch.ones_like(t) for t in unpadded_input_ids]
+        padded_attn_mask = pad_sequence(unpadded_attn_mask, batch_first=True, padding_value=0)
+        padded_input_ids = pad_sequence(unpadded_input_ids, batch_first=True)
+
+        # combine the attention masks
+        full_attn_mask = torch.cat((padded_past_keys_values_attn_mask, padded_attn_mask), dim=1)
+        assert full_attn_mask.shape[1] <= model.config.n_positions, "Presently batches with wide range of prefix and input lengths are not supported due overrun of max model size"
+
+        # make position_ids
+        max_input_len = padded_input_ids.shape[-1]
+        position_ids = torch.stack([torch.arange(cache_length, cache_length + max_input_len) for cache_length in cache_lengths], dim=0)
+        position_ids = position_ids * padded_attn_mask
+        assert (position_ids < model.config.n_positions).all()
+
+        inputs = {
+            'input_ids': padded_input_ids.to(model.device),
+            'past_key_values': padded_past_keys_values,
+            'attention_mask': full_attn_mask.to(model.device),
+            'position_ids': position_ids.to(model.device)
+        }
+
+        return inputs, input_lengths, batch_contexts, batch_continuations
+    
+    def _get_inputs_without_cache(self, batch_of_indices: Sequence[int], cc_pairs: List[Dict[str, Tuple[torch.Tensor, torch.Tensor]]], model: _Model):
+        unpadded_batch = collections.defaultdict(list)
+        input_lengths = []
+        batch_contexts = []
+        batch_continuations = []
+
+        for index in batch_of_indices:
+            for field_name, (context_ids, continuation_ids) in cc_pairs[index].items():
+                ids = torch.cat([context_ids, continuation_ids])[:-1]
+                unpadded_batch[field_name].append(ids)
+
+            input_lengths.append(len(unpadded_batch["input_ids"][-1]))
+            batch_contexts.append(cc_pairs[index]["input_ids"][0])
+            batch_continuations.append(cc_pairs[index]["input_ids"][1])
+
+        padded_batch = {
+            field_name: pad_sequence(tensors, batch_first=True).to(model.device)
+            for field_name, tensors in unpadded_batch.items()
+        }
+        return padded_batch, input_lengths, batch_contexts, batch_continuations